// src/lit.smabu.react/src/types/domain/payment-dto.ts
/* tslint:disable */
/* eslint-disable */
/**
 * LIT.Smabu.API
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * OpenAPI spec version: 1.0
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */

import { Currency } from './currency';
import { CustomerId } from './customer-id';
import { InvoiceId } from './invoice-id';
import { PaymentDirection } from './payment-direction';
import { PaymentId } from './payment-id';
import { PaymentNumber } from './payment-number';
import { PaymentStatus } from './payment-status';
 /**
 * 
 *
 * @export
 * @interface PaymentDTO
 */
export interface PaymentDTO {

    /**
     * @type {PaymentId}
     * @memberof PaymentDTO
     */
    id?: PaymentId;

    /**
     * @type {PaymentNumber}
     * @memberof PaymentDTO
     */
    number?: PaymentNumber;

    /**
     * @type {PaymentDirection}
     * @memberof PaymentDTO
     */
    direction?: PaymentDirection;

    /**
     * @type {Date}
     * @memberof PaymentDTO
     */
    accountingDate?: Date;

    /**
     * @type {string}
     * @memberof PaymentDTO
     */
    details?: string | null;

    /**
     * @type {string}
     * @memberof PaymentDTO
     */
    payer?: string | null;

    /**
     * @type {string}
     * @memberof PaymentDTO
     */
    payee?: string | null;

    /**
     * @type {CustomerId}
     * @memberof PaymentDTO
     */
    customerId?: CustomerId;

    /**
     * @type {InvoiceId}
     * @memberof PaymentDTO
     */
    invoiceId?: InvoiceId;

    /**
     * @type {string}
     * @memberof PaymentDTO
     */
    referenceNr?: string | null;

    /**
     * @type {Date}
     * @memberof PaymentDTO
     */
    referenceDate?: Date | null;

    /**
     * @type {number}
     * @memberof PaymentDTO
     */
    amountDue?: number;

    /**
     * @type {Date}
     * @memberof PaymentDTO
     */
    dueDate?: Date | null;


    /**
     * @type {boolean}
     * @memberof PaymentDTO
     */
    isOverdue?: boolean | null;

    /**
     * @type {number}
     * @memberof PaymentDTO
     */
    amountPaid?: number;

    /**
     * @type {Date}
     * @memberof PaymentDTO
     */
    paidAt?: Date | null;

    /**
     * @type {Currency}
     * @memberof PaymentDTO
     */
    currency?: Currency;

    /**
     * @type {PaymentStatus}
     * @memberof PaymentDTO
     */
    status?: PaymentStatus;

    /**
     * @type {string}
     * @memberof PaymentDTO
     */
    displayName?: string | null;
}